pyproject.toml
[tool.poetry]
name = "oak-signs"
version = "0.1.0"
description = "API for Minecraft notifications management"
authors = ["Tobias <piotr.tobiasz.dev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.75.0"
uvicorn = { extras = ["standard"], version = "^0.17.4" }
structlog = "^21.5.0"
colorama = "^0.4.4"
environs = "^9.5.0"
strawberry-graphql = { extras = ["fastapi"], version = "^0.102.2" }
beanie = "^1.10.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
pre-commit = "^2.17.0"
wemake-python-styleguide = "^0.16.0"
pyproject-flake8 = "^0.0.1-alpha.2"
pytest-asyncio = "^0.18.2"
httpx = "^0.22.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 80

[tool.isort]
line_length = 80
profile = "black"
default_section = "THIRDPARTY"
known_first_party = ["oak_signs"]

[tool.flake8]
exclude = ".git,__pycache__,*/static/*,*/migrations/*,*/test_*/"
max-line-length = 80
inline-quotes = '"'
ignore = "WPS115,WPS306,WPS111,N812,WPS110"
per-file-ignores = """
    */__init__.py:D104
"""

[tool.pytest.ini_options]
testpaths = "tests"
asyncio_mode = "strict"
addopts = "--strict-markers -vv --cache-clear --maxfail=1"
